solved. JSONObjectAdapterGWT behavior change after upgrade

+++ src/main/java/org/sagebionetworks/web/client/widget/entity/renderer/APITableColumnRendererEpochDate.java
@@ -47,8 +47,7 @@ public class APITableColumnRendererEpochDate implements APITableColumnRenderer {
 		for (Iterator iterator2 = colValues.iterator(); iterator2
 				.hasNext();) {
 			String colValue = (String) iterator2.next();
-			//parse as a double, and cast to long (truncate toward zero)
-			Date date = new Date((long)Double.parseDouble(colValue));
+			Date date = new Date(Long.parseLong(colValue));
 			outputValues.add(synapseJSNIUtils.convertDateToSmallString(date));
 		}
 		outputColumnData.put(outputColumnName, outputValues);
